+++ helpers/playlist_manager_menu.js
@@ -943,6 +943,8 @@ function createMenuRightTop() {
 		const options = list.categories();
 		const defOpt = options[0];
 		const optionsLength = options.length;
+		menu.newEntry({menuName: subMenuName, entryText: 'Toogle (click) / Single (Shift + click):', func: null, flags: MF_GRAYED});
+		menu.newEntry({menuName: subMenuName, entryText: 'sep'});
 		menu.newEntry({menuName: subMenuName, entryText: 'Restore all', func: () => {
 			list.filter({categoryState: options});
 		}});
@@ -950,7 +952,13 @@ function createMenuRightTop() {
 		const iInherit = (list.categoryState.length === 1 && list.categoryState[0] !== defOpt ? options.indexOf(list.categoryState[0]) : -1);
 		options.forEach((item, i) => {
 			menu.newEntry({menuName: subMenuName, entryText: item + (i === iInherit ? '\t-inherit-' : ''), func: () => {
-				const categoryState = list.categoryState.indexOf(item) !== -1 ? list.categoryState.filter((categ) => {return categ !== item;}) : (item === defOpt ? [defOpt, ...list.categoryState] : list.categoryState.concat([item]).sort());
+				let categoryState;
+				// Disable all other tags when pressing shift
+				if (utils.IsKeyPressed(VK_SHIFT)) {
+					categoryState = [item];
+				} else {
+					categoryState = list.categoryState.indexOf(item) !== -1 ? list.categoryState.filter((categ) => {return categ !== item;}) : (item === defOpt ? [defOpt, ...list.categoryState] : list.categoryState.concat([item]).sort());
+				}
 				list.filter({categoryState});
 			}});
 			menu.newCheckMenu(subMenuName, item, void(0), () => {return list.categoryState.indexOf(item) !== -1;});
@@ -961,6 +969,8 @@ function createMenuRightTop() {
 		const options = list.tags();
 		const defOpt = options[0];
 		const optionsLength = options.length;
+		menu.newEntry({menuName: subMenuName, entryText: 'Toogle (click) / Single (Shift + click):', func: null, flags: MF_GRAYED});
+		menu.newEntry({menuName: subMenuName, entryText: 'sep'});
 		menu.newEntry({menuName: subMenuName, entryText: 'Restore all', func: () => {
 			list.filter({tagState: options});
 		}});
@@ -968,7 +978,13 @@ function createMenuRightTop() {
 		const bInherit = list.tagState.indexOf(defOpt) === -1;
 		options.forEach((item, i) => {
 			menu.newEntry({menuName: subMenuName, entryText: item + (bInherit && i !== 0 ? '\t-inherit-' : ''), func: () => {
-				const tagState = list.tagState.indexOf(item) !== -1 ? list.tagState.filter((tag) => {return tag !== item;}) : (item === defOpt ? [defOpt, ...list.tagState] : list.tagState.concat([item]).sort());
+				let tagState;
+				// Disable all other categories when pressing shift
+				if (utils.IsKeyPressed(VK_SHIFT)) {
+					tagState = [item];
+				} else {
+					tagState = list.tagState.indexOf(item) !== -1 ? list.tagState.filter((tag) => {return tag !== item;}) : (item === defOpt ? [defOpt, ...list.tagState] : list.tagState.concat([item]).sort());
+				}
 				list.filter({tagState});
 			}});
 			menu.newCheckMenu(subMenuName, item, void(0), () => {return list.tagState.indexOf(item) !== -1;});